fix: move to component to avoid breaking react rules of hooks
